+++ kong/plugins/openid-connect/handler.lua
@@ -7,7 +7,7 @@
 
 local OICHandler = {
   PRIORITY = 1000,
-  VERSION  = "2.0.0",
+  VERSION  = "2.0.1",
 }